moved view into own dir
